WIP: Fixes   [skip ci]

+++ apps/api/src/app/topics/use-cases/add-bulk-subscribers/add-bulk-subscribers.use-case.ts
@@ -55,14 +55,15 @@ export class AddBulkSubscribersUseCase {
       results.push({ key: topicKey, existingExternalSubscribers, nonExistingExternalSubscribers });
     }
 
-    for (const topicKey of command.removeKeys) {
-      await this.topicSubscribersRepository.removeSubscribers(
-        command.environmentId,
-        command.organizationId,
-        topicKey,
-        command.subscribers
-      );
-    }
+    if (command.removeKeys)
+      for (const topicKey of command.removeKeys) {
+        await this.topicSubscribersRepository.removeSubscribers(
+          command.environmentId,
+          command.organizationId,
+          topicKey,
+          command.subscribers
+        );
+      }
 
     return results;
   }
